requirements.txt
fastapi
uvicorn
streamlit
sqlalchemy
pydantic
python-dotenv
httpx
requests
python-multipart
google-generativeai

